Make overload of get_mapped_byte() that takes size_t index a #1 overload.

+++ src/SegMap595.cpp
@@ -184,7 +184,7 @@ int32_t SegMap595Class::get_status()
     return _status;
 }
 
-uint8_t SegMap595Class::get_mapped_byte(uint32_t index)
+uint8_t SegMap595Class::get_mapped_byte(size_t index)
 {
     if (_status < 0 || index >= _glyph_set_chosen->glyph_num) {
         return 0;
@@ -193,9 +193,9 @@ uint8_t SegMap595Class::get_mapped_byte(uint32_t index)
     return _mapped_bytes[index];
 }
 
-uint8_t SegMap595Class::get_mapped_byte(size_t index)
+uint8_t SegMap595Class::get_mapped_byte(uint32_t index)
 {
-    return get_mapped_byte(static_cast<uint32_t>(index));
+    return get_mapped_byte(static_cast<size_t>(index));
 }
 
 uint8_t SegMap595Class::get_mapped_byte(unsigned char char_represented)
